Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-snap/snap-after-relayout/remove-current-target-expected.txt


FAIL Removing the current snap target should make the scroller snap to a new target. assert_equals: expected 100 but got 0
